Move password actions to new password.js
